# perf: speed up the first loading of namespace when startup meet 404