--- a/apollo-client/pom.xml
+++ b/apollo-client/pom.xml
@@ -92,6 +92,11 @@
 			<artifactId>log4j-core</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.mock-server</groupId>
+			<artifactId>mockserver-netty</artifactId>
+			<scope>test</scope>
+		</dependency>
 		<!-- end of test -->
 	</dependencies>
 </project>