--- a/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/AbstractConfigRepository.java
+++ b/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/AbstractConfigRepository.java
@@ -73,4 +73,9 @@ public abstract class AbstractConfigRepository implements ConfigRepository {
       }
     }
   }
+
+  @Override
+  public void initialize() {
+    this.sync();
+  }
 }