--- a/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/ConfigRepository.java
+++ b/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/ConfigRepository.java
@@ -53,4 +53,9 @@ public interface ConfigRepository {
    * @return the config's source type
    */
   ConfigSourceType getSourceType();
+
+  /**
+   * Initialize the repository.
+   */
+  void initialize();
 }