--- a/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/ConfigRepository.java
+++ b/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/ConfigRepository.java
@@ -57,5 +57,5 @@ public interface ConfigRepository {
   /**
    * Initialize the repository.
    */
-  void initialize();
+  default void initialize() {}
 }