--- a/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/LocalFileConfigRepository.java
+++ b/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/LocalFileConfigRepository.java
@@ -71,7 +71,6 @@ public class LocalFileConfigRepository extends AbstractConfigRepository
     m_configUtil = ApolloInjector.getInstance(ConfigUtil.class);
     this.setLocalCacheDir(findLocalCacheDir(), false);
     this.setUpstreamRepository(upstream);
-    // this.trySync();
   }
 
   void setLocalCacheDir(File baseDir, boolean syncImmediately) {