--- a/apollo-client/src/test/java/com/ctrip/framework/apollo/BaseIntegrationTest.java
+++ b/apollo-client/src/test/java/com/ctrip/framework/apollo/BaseIntegrationTest.java
@@ -18,34 +18,19 @@ package com.ctrip.framework.apollo;
 
 import com.ctrip.framework.apollo.core.ConfigConsts;
 import com.ctrip.framework.apollo.core.MetaDomainConsts;
+import com.ctrip.framework.apollo.core.dto.ApolloConfig;
+import com.ctrip.framework.apollo.core.dto.ServiceDTO;
 import java.io.IOException;
 import java.net.ServerSocket;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.eclipse.jetty.server.Request;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.handler.AbstractHandler;
-import org.eclipse.jetty.server.handler.ContextHandler;
-import org.eclipse.jetty.server.handler.ContextHandlerCollection;
-import org.junit.After;
-import org.junit.Before;
 
 import com.ctrip.framework.apollo.build.MockInjector;
-import com.ctrip.framework.apollo.core.dto.ServiceDTO;
 import com.ctrip.framework.apollo.core.enums.Env;
 import com.ctrip.framework.apollo.core.utils.ClassLoaderUtil;
 import com.ctrip.framework.apollo.util.ConfigUtil;
-import com.google.common.collect.Lists;
 import com.google.gson.Gson;
-import org.junit.Rule;
-import org.junit.rules.TestRule;
-import org.junit.rules.TestWatcher;
-import org.junit.runner.Description;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.test.util.ReflectionTestUtils;
@@ -67,25 +52,51 @@ public abstract class BaseIntegrationTest {
   protected static int refreshInterval;
   protected static TimeUnit refreshTimeUnit;
   protected static boolean propertiesOrderEnabled;
-  private Server server;
+  private MockedConfigService mockedConfigService;
   protected Gson gson = new Gson();
 
-  @Rule
-  public TestRule watcher = new TestWatcher() {
-    protected void starting(Description description) {
-      logger.info("Starting test: " + description.getMethodName());
-    }
+  protected MockedConfigService newMockedConfigService() {
+    this.mockedConfigService = new MockedConfigService(port);
+    this.mockedConfigService.init();
+    this.mockMetaServer();
+    return this.mockedConfigService;
+  }
 
-    protected void finished(Description description) {
-      logger.info("Finished test: " + description.getMethodName());
-    }
-  };
+  protected void mockMetaServer() {
+    mockMetaServer(false);
+  }
 
-  @Before
+  protected void mockMetaServer(boolean failedAtFirstTime) {
+    final ServiceDTO someServiceDTO = new ServiceDTO();
+    someServiceDTO.setAppName(someAppName);
+    someServiceDTO.setInstanceId(someInstanceId);
+    someServiceDTO.setHomepageUrl(configServiceURL);
+    this.mockedConfigService.mockMetaServer(failedAtFirstTime, someServiceDTO);
+  }
+
+  public void mockConfigs(
+      int mockedStatusCode,
+      ApolloConfig apolloConfig
+  ) {
+    this.mockConfigs(false, mockedStatusCode, apolloConfig);
+  }
+
+  public void mockConfigs(
+      boolean failedAtFirstTime,
+      int mockedStatusCode,
+      ApolloConfig apolloConfig
+  ) {
+    this.mockedConfigService.mockConfigs(
+        failedAtFirstTime, mockedStatusCode, apolloConfig
+    );
+  }
+
+  @BeforeEach
   public void setUp() throws Exception {
     someAppId = "1003171";
     someClusterName = "someClusterName";
     someDataCenter = "someDC";
+
     refreshInterval = 5;
     refreshTimeUnit = TimeUnit.MINUTES;
     propertiesOrderEnabled = false;
@@ -99,7 +110,7 @@ public abstract class BaseIntegrationTest {
     MockInjector.setInstance(ConfigUtil.class, new MockConfigUtil());
   }
 
-  @After
+  @AfterEach
   public void tearDown() throws Exception {
     //as ConfigService is singleton, so we must manually clear its container
     ConfigService.reset();
@@ -107,61 +118,10 @@ public abstract class BaseIntegrationTest {
     System.clearProperty(ConfigConsts.APOLLO_META_KEY);
     ReflectionTestUtils.invokeMethod(MetaDomainConsts.class, "reset");
 
-    if (server != null && server.isStarted()) {
-      server.stop();
+    if (mockedConfigService != null) {
+      mockedConfigService.close();
+      mockedConfigService = null;
     }
-  }
-
-  /**
-   * init and start a jetty server, remember to call server.stop when the task is finished
-   *
-   * @param handlers
-   * @throws Exception
-   */
-  protected Server startServerWithHandlers(ContextHandler... handlers) throws Exception {
-    server = new Server(port);
-
-    ContextHandlerCollection contexts = new ContextHandlerCollection();
-    contexts.setHandlers(handlers);
-    contexts.addHandler(mockMetaServerHandler());
-
-    server.setHandler(contexts);
-    server.start();
-
-    return server;
-  }
-
-  protected ContextHandler mockMetaServerHandler() {
-    return mockMetaServerHandler(false);
-  }
-
-  protected ContextHandler mockMetaServerHandler(final boolean failedAtFirstTime) {
-    final ServiceDTO someServiceDTO = new ServiceDTO();
-    someServiceDTO.setAppName(someAppName);
-    someServiceDTO.setInstanceId(someInstanceId);
-    someServiceDTO.setHomepageUrl(configServiceURL);
-    final AtomicInteger counter = new AtomicInteger(0);
-
-    ContextHandler context = new ContextHandler("/services/config");
-    context.setHandler(new AbstractHandler() {
-      @Override
-      public void handle(String target, Request baseRequest, HttpServletRequest request,
-          HttpServletResponse response) throws IOException, ServletException {
-        if (failedAtFirstTime && counter.incrementAndGet() == 1) {
-          response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-          baseRequest.setHandled(true);
-          return;
-        }
-        response.setContentType("application/json;charset=UTF-8");
-        response.setStatus(HttpServletResponse.SC_OK);
-
-        response.getWriter().println(gson.toJson(Lists.newArrayList(someServiceDTO)));
-
-        baseRequest.setHandled(true);
-      }
-    });
-
-    return context;
   }
 
   protected void setRefreshInterval(int refreshInterval) {