--- a/apollo-client/src/test/java/com/ctrip/framework/apollo/internals/LocalFileConfigRepositoryTest.java
+++ b/apollo-client/src/test/java/com/ctrip/framework/apollo/internals/LocalFileConfigRepositoryTest.java
@@ -186,6 +186,8 @@ public class LocalFileConfigRepositoryTest {
     LocalFileConfigRepository localFileConfigRepository =
         new LocalFileConfigRepository(someNamespace, upstreamRepo);
 
+    assertEquals(ConfigSourceType.LOCAL, localFileConfigRepository.getSourceType());
+    localFileConfigRepository.trySync();
     assertEquals(someSourceType, localFileConfigRepository.getSourceType());
 
     localFileConfigRepository.setLocalCacheDir(someBaseDir, true);