## What's the purpose of this PR

Speed up the user's startup time,

by decrease the number of http requests.

## Which issue(s) this PR fixes:
Fixes https://github.com/apolloconfig/apollo-java/issues/60

## Brief changelog

before change, it use 11494 ms to load 10 namespace when meet 404, 
[before-change.log](https://github.com/apolloconfig/apollo-java/files/15237228/before-change.log)

after delete 2 http requests when context init, it use 1490 ms.
[after-change.log](https://github.com/apolloconfig/apollo-java/files/15237229/after-change.log)

i.e **from 11494 ms to 1490 ms**.

the code in 
https://github.com/apolloconfig/apollo-java/blob/bcc40531f32e74d0c639b5e5d2ab2995fd00c3d6/apollo-client/src/main/java/com/ctrip/framework/apollo/internals/DefaultConfig.java#L69
will ensure pull at least one time from remote when startup.

Follow this checklist to help us incorporate your contribution quickly and easily:

- [x] Read the [Contributing Guide](https://github.com/apolloconfig/apollo/blob/master/CONTRIBUTING.md) before making this pull request.
- [x] Write a pull request description that is detailed enough to understand what the pull request does, how, and why.
- [x] Write necessary unit tests to verify the code.
- [x] Run `mvn clean test` to make sure this pull request doesn't break anything.
- [x] Update the [`CHANGES` log](https://github.com/apolloconfig/apollo-java/blob/master/CHANGES.md).


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Introduced a new initialization process in the configuration repository, enhancing the setup mechanism.

- **Bug Fixes**
  - Adjusted initialization calls to ensure proper configuration setup without redundant sync calls.

- **Tests**
  - Added assertions in tests to verify the source type before and after initialization, improving test coverage and reliability.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->